+++ lib/ui-text/index.ts
@@ -1,3 +1,4 @@
+export { default as UIHeading } from "./ui-heading";
 export { default as UILabel } from "./ui-label";
 export { default as UIPrimary } from "./ui-primary";
 export { default as UISecondary } from "./ui-secondary";