+++ lib/ui-text/ui-secondary.ts
@@ -1,4 +1,3 @@
-import { html } from "lit";
 import { customElement } from "lit/decorators.js";
 import UIText from "./ui-text";
 
@@ -10,13 +9,8 @@ import UIText from "./ui-text";
 @customElement("ui-secondary")
 class UISecondary extends UIText {
     size: string = "0.9rem";
-    casl: number = 1;
-    mono: number = 0;
+    wght: number = 350;
     slnt: number = -15;
-
-    protected render() {
-        return html`${super.render()}`;
-    }
 }
 
 export default UISecondary;