+++ lib/ui-text/ui-text.ts
@@ -5,22 +5,24 @@ import { customElement, property } from "lit/decorators.js";
  * **Tag**: `ui-text`
  *
  * **Attributes**:
- *  - casl: `number`
- *  - mono: `number`
- *  - slnt: `number`
- *  - size: `string`
  *  - family: `string`
+ *  - size: `string`
+ *  - mono: `number`
+ *  - casl: `number`
+ *  - wght: `number`
+ *  - slnt: `number`
+ *  - crsv: `number`
  *
  * **Slots**:
  *  - ""
  */
 @customElement("ui-text")
 class UIText extends LitElement {
-    /**
-     * Range between 0 - 1
-     */
-    @property({ type: Number, attribute: "casl", reflect: true })
-    casl: number = 1;
+    @property({ type: Number, attribute: "family", reflect: true })
+    family: string = "var(--ui-fontFamily)";
+
+    @property({ type: Number, attribute: "size", reflect: true })
+    size: string = "var(--ui-fontSize)";
 
     /**
      * Range between 0 - 1
@@ -29,16 +31,25 @@ class UIText extends LitElement {
     mono: number = 0;
 
     /**
+     * Range between 0 - 1
+     */
+    @property({ type: Number, attribute: "casl", reflect: true })
+    casl: number = 1;
+
+    @property({ type: Number, attribute: "wght", reflect: true })
+    wght: number = 400;
+
+    /**
      * Range between -15 - 0
      */
     @property({ type: Number, attribute: "slnt", reflect: true })
     slnt: number = 0;
 
-    @property({ type: Number, attribute: "size", reflect: true })
-    size: string = "var(--ui-fontSize)";
-
-    @property({ type: Number, attribute: "family", reflect: true })
-    family: string = "var(--ui-fontFamily)";
+    /**
+     * Range between 0 / 0.5 / 1
+     */
+    @property({ type: Number, attribute: "crsv", reflect: true })
+    crsv: number = 0.5;
 
     static get styles() {
         return css`
@@ -46,40 +57,28 @@ class UIText extends LitElement {
                 display: inline-block;
                 overflow-wrap: anywhere;
 
-                font-size: var(--ui-text_size, var(--ui-fontSize));
-                font-family: var(--ui-text_family, var(--ui-fontFamily));
+                font-family: var(--_family) !important;
+                font-size: var(--_size) !important;
                 font-variation-settings:
-                    "CASL" var(--ui-text_casl, 1),
-                    "MONO" var(--ui-text_mono, 0),
-                    "slnt" var(--ui-text_slnt, 0);
+                    "MONO" var(--_mono),
+                    "CASL" var(--_casl),
+                    "wght" var(--_wght),
+                    "slnt" var(--_slnt),
+                    "CRSV" var(--_crsv) !important;
             }
         `;
     }
 
     protected render() {
+        this.style.setProperty("--_family", this.family);
+        this.style.setProperty("--_size", this.size);
+        this.style.setProperty("--_mono", this.mono.toString());
+        this.style.setProperty("--_casl", this.casl.toString());
+        this.style.setProperty("--_wght", this.wght.toString());
+        this.style.setProperty("--_slnt", this.slnt.toString());
+        this.style.setProperty("--_crsv", this.crsv.toString());
+
         return html`<slot></slot>`;
-    }
-
-    attributeChangedCallback(
-        name: string,
-        _old: string | null,
-        value: string | null,
-    ): void {
-        super.attributeChangedCallback(name, _old, value);
-
-        switch (name) {
-            case "casl":
-            case "mono":
-            case "slnt":
-            case "size":
-            case "family":
-                if (value === null) {
-                    this.style.removeProperty(`--ui-text_${name}`);
-                    return;
-                }
-                this.style.setProperty(`--ui-text_${name}`, value);
-                break;
-        }
     }
 }